pyproject.toml
[tool.poetry]
name = "weekly"
version = "0.1.33"
description = "A comprehensive Python project quality analyzer that provides actionable next steps for improving your project"
authors = ["Tom Sapletta <info@softreck.dev>"]
license = "Apache-2.0"
readme = "README.md"
repository = "https://github.com/wronai/weekly"
documentation = "https://weekly.readthedocs.io"
keywords = [
    "code quality", 
    "analysis", 
    "linter", 
    "formatter", 
    "ci-cd", 
    "dependencies",
    "static analysis",
    "code review",
    "best practices"
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development",
    "Topic :: Software Development :: Quality Assurance",
    "Topic :: Software Development :: Testing",
    "Topic :: Software Development :: Build Tools",
    "Topic :: Utilities",
    "Typing :: Typed",
]

[tool.poetry.dependencies]
python = ">=3.8.1,<4.0"
click = "^8.1.3"
tomli = { version = "^2.0.1", python = "<3.11" }
toml = "^0.10.2"  # For parsing TOML files
pyyaml = "^6.0.0"
rich = "^13.4.1"  # For beautiful console output
python-dotenv = "^1.0.0"  # For .env file support

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-cov = "^4.1.0"
pytest-mock = "^3.11.1"
pytest-xdist = "^3.3.1"
black = "^23.7.0"
isort = "^5.12.0"
mypy = "^1.5.0"
flake8 = "^6.1.0"
pre-commit = "^3.3.3"
pip-audit = "^2.6.1"
coverage = {extras = ["toml"], version = "^7.2.7"}

[tool.poetry.scripts]
weekly = "weekly.cli:main"

[tool.poetry.urls]
"Homepage" = "https://github.com/wronai/weekly"
"Bug Tracker" = "https://github.com/wronai/weekly/issues"
"Documentation" = "https://weekly.readthedocs.io"
"Changelog" = "https://github.com/wronai/weekly/blob/main/CHANGELOG.md"
"Source Code" = "https://github.com/wronai/weekly"
"Discussions" = "https://github.com/wronai/weekly/discussions"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py38']
include = '\.pyi?$'

[tool.isort]
profile = "black"
line_length = 88
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = false
use_parentheses = true
ensure_newline_before_comments = true
known_first_party = ["weekly"]

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_optional = true
strict_equality = true
disallow_any_generics = true
disallow_subclassing_any = true
disallow_untyped_calls = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --cov=weekly --cov-report=term-missing --cov-report=xml"
pythonpath = ["."]

[coverage.run]
source = ["weekly"]
omit = [
    "**/__init__.py",
    "**/__main__.py",
    "**/version.py",
    "**/tests/**",
]

[coverage.report]
show_missing = true
skip_covered = true
fail_under = 80

[coverage.html]
directory = "htmlcov"

[coverage.xml]
output = "coverage.xml"

[tool.ruff]
line-length = 88
target-version = "py38"
select = [
    "E", "W", "F", "I", "B", "C4", "ISC", "ICN", "G", "INP", 
    "PIE", "T20", "A", "COM", "C90", "N", "UP", "YTT", "FBT", "Q", "RUF"
]
ignore = ["E501", "W503", "E203"]

[tool.ruff.per-file-ignores]
"tests/*" = ["S101"]

[metadata]
license = "Apache-2.0"
license_files = ["LICENSE", "NOTICE"]
